### 2021-12-01 #3970 @whedon
[PRE REVIEW]: QDistRnd: A GAP package for computing the distance of quantum error-correcting codes
**Submitting author:** @LeonidPryadko (<a href="http://orcid.org/0000-0002-4990-0259">Leonid P. Pryadko</a>)
**Repository:** <a href="https://github.com/QEC-pages/QDistRnd/" target ="_blank">https://github.com/QEC-pages/QDistRnd/</a>
**Version:** 0.7.0
