+++ joss/QDistRnd.bib
@@ -140,7 +140,7 @@
                    Computer Science). Vol. 5557 (2009), pp. 259-273],
                    who asked whether homological codes with such
                    parameters could exist at all.},
-  doi =           {http://dx.doi.org/10.1063/1.4891487},
+  doi =           {10.1063/1.4891487},
   url =           {http://scitation.aip.org/content/aip/journal/jmp/55/8/
                   10.1063/1.4891487},
 }
