### 2021-12-01 #3970 @LeonidPryadko
How do I add a license?
I already have a license listed in documentation (the manual), it says:
Copyright

© 2021 by Leonid P. Pryadko and Vadim A. Shabashov

*QDistRnd* package is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License
<http://www.fsf.org/licenses/gpl.html> as published by the Free Software
Foundation; either version 2 of the License, or (at your option) any later
version.

Thanks,

Leonid

On Wed, Dec 1, 2021 at 3:19 PM Kristen Thyng ***@***.***>
wrote:

> @LeonidPryadko <https://github.com/LeonidPryadko> Can you add a license
> to your repo?
>
