### 2021-12-01 #3970 @kthyng
You can get more info [here](https://joss.readthedocs.io/en/latest/submitting.html#typical-paper-submission-flow). In short you need to have a file called LICENSE that meets our guidelines.
