### 2021-12-01 #3970 @LeonidPryadko
I just added GPL 2.0 (in the documentation, it says GPL 2.0 or any later
version).

Will it do?
Thanks,

Leonid

P.S.
I also corrected a doi in one of the references.

On Wed, Dec 1, 2021 at 3:28 PM Kristen Thyng ***@***.***>
wrote:

> You can get more info here
> <https://joss.readthedocs.io/en/latest/submitting.html#typical-paper-submission-flow>.
> In short you need to have a file called LICENSE that meets our guidelines.
>
