### 2021-12-01 #3970 @kthyng
Hm I thought that command would re-check for the license. Anyway it looks ok.
