### 2021-12-08 #3970 @danielskatz
👋 @LeonidPryadko - Can you suggest any potential reviewers? If so, please mention them here in this thread (without tagging them with an @). [This list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list), but suggestions from elsewhere are also welcome.
