### 2021-12-09 #3970 @LeonidPryadko
Dear Daniel:

As for reviewers, I may suggest Cristophe Vuillot <
***@***.***> and Nikolas Breuckmann ***@***.***> ;
these people have used GAP and also know about quantum error correcting
codes.

Regarding the README, the distribution already has a short README.md file.
Would you like me to rename it to just README, or add details?

Thanks,

Leonid

On Thu, Dec 9, 2021 at 9:33 AM Daniel S. Katz ***@***.***>
wrote:

> @LeonidPryadko <https://github.com/LeonidPryadko> - You also need a
> create a README in your repo.
>
