### 2021-12-09 #3970 @danielskatz
@LeonidPryadko - as you can see on https://github.com/openjournals/joss-reviews/issues/3970#issuecomment-990104679, those emails didn't come across in a useful way. Can you type them into this thread, rather than replying by email?

And the README I see is https://github.com/QEC-pages/QDistRnd/blob/main/README.md  - this doesn't meet what JOSS requires in https://joss.readthedocs.io/en/latest/review_criteria.html#documentation
