### 2022-01-03 #3970 @danielskatz
@LeonidPryadko - I've been on vacation the last two weeks and will still be this week, but we do need to find reviewers, and for you to create a README that meets the JOSS requirements.
