### 2022-01-06 #3970 @LeonidPryadko
Here are the email addresses of the suggested reviewers:
 Cristophe Vuillot <
christophe.vuillot@inria.fr > and Nikolas Breuckmann < nikobreu@gmail.com >

I extended the README.md file for the package distribution at github; hope it contains enough information now.  Please let me know if any additional information needed is still missing.
