### 2022-01-13 #3970 @danielskatz
👋 @dimpase, @fingolfin, @ChrisJefferson, @ruthhoffmann - You were suggested to me as potential reviewers of this JOSS submission by Alexander Konovalov...

Would one of you be willing to review it? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
