### 2022-01-14 #3970 @dimpase
I'd like to suggest to run this by a "real" quantum codes expert, e.g. @fredezerman (he might need a ping by email: https://personal.ntu.edu.sg/fredezerman/Fred_CV.pdf)

I find the package description lacking in the theoretical sense - it seems that the only detailed description of what's done there is in the code itself.
