### 2022-01-14 #3970 @danielskatz
Thanks @dimpase - I'll check with @fredezerman.

Re your other comment, can you be more specific about what you mean by "the package description"?  Note that JOSS papers are intended to be fairly short, though the repository (README, documentation, etc.) should be at an appropriate level of detail.
