### 2022-01-14 #3970 @danielskatz
👋 @fredezerman - As you can see above, you were suggested to me as a potential reviewer of this JOSS submission by @dimpase...

Would one of you be willing to review it? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I'll also be sending you an email about this shortly, in case that's a better way to get you
