### 2022-01-14 #3970 @dimpase
> what you mean by "the package description"

Basically, its manual says
```
The GAP package QDistRnd gives a reference implementation of a probabilistic algorithm for finding the distance of a q-ary quantum low-density parity-check code linear over a finite field F=GF(q).
```
It appears that the only description of the algorithm is in the manual/code itself. IMHO it's not what "reference implementation" is meant to be - there seems to be no other published material on it, or other implementations of it.
