### 2022-01-14 #3970 @LeonidPryadko
A few comments in relation to the comments of @dimpase.

1. I should remark that the manual indeed gives a detailed description of the algorithm; we thought it unnecessary to duplicate the information in the README file given that it is rather technical.

2. If deemed necessary, I will be happy to add the following two paragraphs describing the algorithm to the README file:

In the case of  a CSS code with two stabilizer generator matrices, $H_x$ and $H_z$ with mutually orthogonal rows, the algorithm amounts to the following steps:
  * Construct matrix $G_z$ dual to $H_x$ (mutually orthogonal rows and complementary ranks);
  * Permute columns of $G_z$ randomly and perform row operations (Gauss elimination) to get a reduced row echelon form, followed by the inverse column permutation;
  * Compare weights of the rows of thus obtained modified $G_z$ which are linearly independent from the rows of $H_z$; the minimum gives an upper bound for the distance $d_z$.
  * A minimum weight code word can be found after repeating these steps for a sufficient number of times.

The algorithm works because it draws random code words from a distribution biased to lower weights.  It is generic, meaning that it works with any code.  It is also a generalization of a well established "information set" class of algorithms for classical linear codes (see references in the program manual), also used in cryptography.  In all cases, the complexity is roughly exponential in the code distance.  Unfortunately, there is no accurate estimate of the complexity exponent which could be used as a stopping criterion, definitely so in the case of quantum LDPC codes.

3. The algorithm in the package is a bit different but is related to that described in Dumer, Kovalev, and Pryadko "Distance verification for classical and quantum LDPC codes", IEEE Trans. Inf. Th. **63**, 4675 (2017), where the exponent of question in the case of a family of random *classical* LDPC codes was related to the corresponding erasure threshold.  Because of this similarity, the relation to the well-established information set algorithms, and also since there is no performance guarantee, we decided that it does not deserve a regular independent publication.

4. Yet, it turned very useful in my own research (both published papers and some work in progress).  In addition, for a number of years I've been reviewing manuscripts (mostly submitted to Quantum Information & Computation) describing new constructions of q-ary quantum codes, often with lower distance bounds but no actual numerics to verify the distances.  I kept writing in my reports that in this day and age, the distances should be actually computed, but could not suggest a package to do it.   For these reasons, I decided to spend the time and get a student to help prepare the GAP package.

5. The GAP package `QDistRnd`, at least in the present version, is substantially slower than the C program which uses `m4ri` library for linear algebra.  On the other hand, it works with q-ary codes. This is why we say in the manual that it gives a 'reference' implementation, i.e., not heavily optimized for performance.

6. Among coding theory specialists, both quantum and classical, who are familiar with the algorithm (but not the GAP package) are Pavel Panteleev <panteleev@intsys.msu.ru> and Gleb Kalachev <kalachev.gleb@huawei.com>.  Some time ago I shared with them a C program for calculating the distance of binary CSS codes which uses a version of the same algorithm, and they have tried it.  We even had a discussion about the performance of the algorithm and that it would be nice to get an accurate estimate of the exponent for specific classes of quantum codes.

I am not sure, though, whether either of these people have ever used GAP; this is why I did not mention the names previously.

Hope it helps,

Leonid
