### 2022-02-01 #4120 @whedon
[REVIEW]: QDistRnd: A GAP package for computing the distance of quantum error-correcting codes
**Submitting author:** @LeonidPryadko (<a href="http://orcid.org/0000-0002-4990-0259">Leonid P. Pryadko</a>)
**Repository:** https://github.com/QEC-pages/QDistRnd/
**Branch with paper.md** (empty if default branch):
**Version:** 0.8.5
