### 2022-02-02 #4120 @danielskatz
@pan-pavel - I saw you tried to open a couple of issues using the GitHub Convert to issue functionality when working through this checklist. Instead, please open any new issues associated with your review in the [software repository associated with the submission](https://github.com/QEC-pages/QDistRnd/) and mention `openjournals/joss-reviews#4120` so that a link is created to this thread (and I can keep an eye on what is happening)
