### 2022-02-02 #4120 @pan-pavel
> I've done this for you in these cases, as you can see above
>
Thanks a lot, Daniel!

> 2 февр. 2022 г., в 04:07, Daniel S. Katz ***@***.***> написал(а):
>
> ﻿
> I've done this for you in these cases, as you can see above
>
