### 2022-02-02 QEC-pages/QDistRnd#5 @pan-pavel
Major issue with the representation of field elements in the MTXE format
From [openjournals/joss-reviews#4120](https://github.com/openjournals/joss-reviews/issues/4120)

Subsection 5.1-1 (Representation of field elements via integers) of the documentation says that:

> In the actual file format, we chose to represent non-zero elements as integers, specifying the powers of a primitive element, while the zero field element is represented as -1. Notice that, depending on the choice of the primitive element, the elements of the field may be permuted. However, any such permutation is a field isomorphism preserving the multiplication table. In particular, orthogonality between the rows of the matrices is necessarily preserved.

However, unfortunately, the statement "any such permutation is a field isomorphism" is incorrect. Indeed, if we replace a primitive element _a_ with another primitive element _a_^i (a root of a different primitive polynomial), then the map _x_ -> _x_^_i_ is **not** a field isomorphism since it does not preserve the additive structure, i.e., (_x_ + _y_)^_i_ is not always equal to _x_^_i_ + _y_^_i_. In short, if we change the primitive element, then the multiplication table is preserved, while the addition table is not.  Please, check this [link](https://math.stackexchange.com/questions/40326/primitive-polynomials-of-finite-fields) for a discussion of such issues.

Hence, I strongly suggest adding into the proposed MTXE format a parameter for the primitive element of GF(_p_^_s_) when _s_ > 1.

I also think that in Subsection 5.2 (Example MTXE files), an example of a matrix over GF(_p_^_s_) where _s_ > 1 should be given. Otherwise, it is not very clear how to use the MTXE format in such cases.
