+++ joss/paper.md
@@ -30,7 +30,7 @@ bibliography: QDistRnd.bib
 # Summary
 
 The GAP package `QDistRnd` implements a probabilistic algorithm for
-finding the minimum distance of a quantum code linear over a finite
+finding the minimum distance of a quantum low-density parity-check code linear over a finite
 field $\mathop{\rm GF}(q)$. At each step several codewords are
 randomly drawn from a distribution biased toward smaller weights. The
 corresponding weights are used to update the upper bound on the
