### 2022-02-07 #4120 @danielskatz
👋 @oscarhiggott and @pan-pavel - this is just a quick check-in here a week into the process.

@oscarhiggott, it looks like you haven't yet accepted [the GitHub invitation](https://github.com/openjournals/joss-reviews/invitations) to the review process. Please do when you have a minute, as described in the first two comments in this issue.

@pan-pavel - thanks for starting your review, checking-off items, and adding issues where you've found problems.

@LeonidPryadko - thanks for responding to some of the issues so far.
