### 2022-02-07 QEC-pages/QDistRnd#5 @LeonidPryadko
TODO:
* add primitive polynomial (PP) to the format description
* update `ReadMTXE` and `WriteMTXE` accordingly, including the actions/error messages when PP is not defined
* verification and field element conversion routine in case a different polynomial is used internally
* an option to save with a primitive element other than the one used internally (?)
