### 2022-02-17 QEC-pages/QDistRnd#9 @oscarhiggott
Tests failing
From openjournals/joss-reviews#4120

I've followed the instructions in the README to install and test the package and get the errors below when running the tests (below is on a mac but I get similar errors with linux).

```
 ┌───────┐   GAP 4.11.1 of 2021-03-02
 │  GAP  │   https://www.gap-system.org
 └───────┘   Architecture: arm-apple-darwin21.3.0-default64-kv7
 Configuration:  gmp 6.2.1, GASMAN
 Loading the library LoadPackage("QDistRnd");and packages ...
 Packages:   AClib 1.3.2, Alnuth 3.1.2, AtlasRep 2.1.0,
             AutoDoc 2020.08.11, AutPGrp 1.10.2, Browse 1.8.11,
             CaratInterface 2.3.3, CRISP 1.4.5, Cryst 4.1.23,
             CrystCat 1.1.9, CTblLib 1.3.1, FactInt 1.6.3, FGA 1.4.0,
             Forms 1.2.5, GAPDoc 1.6.4, genss 1.6.6, IO 4.7.0,
             IRREDSOL 1.4.1, LAGUNA 3.9.3, orb 4.8.3, Polenta 1.3.9,
             Polycyclic 2.16, PrimGrp 3.4.1, RadiRoot 2.8, recog 1.3.2,
             ResClasses 4.7.2, SmallGrp 1.4.2, Sophus 1.24, SpinSym 1.5.2,
             TomLib 1.2.9, TransGrp 3.0, utils 0.69
 Try '??help' for help. See also '?copyright', '?cite' and '?authors'
gap> LoadPackage("QDistRnd");
#I  You may wish to install the xgap package
#I  and enjoy the graphic capabilities of SONATA.

  ___________________________________________________________________________
 /        ___
||       /   \                 /\    Version 2.9.1
||      ||   ||  |\    |      /  \               /\       Erhard Aichinger
 \___   ||   ||  |\\   |     /____\_____________/__\      Franz Binder
     \  ||   ||  | \\  |    /      \     ||    /    \     Juergen Ecker
     ||  \___/   |  \\ |   /        \    ||   /      \    Peter Mayr
     ||          |   \\|  /          \   ||               Christof Noebauer
 \___/           |    \|                 ||

 System    Of   Nearrings     And      Their Applications
 Info: https://gap-packages.github.io/sonata/

   ____                          |
  /            \           /   --+--  Version 3.15
 /      |    | |\\        //|    |
|    _  |    | | \\      // |     the GUAVA Group
|     \ |    | |--\\    //--|
 \     ||    | |   \\  //   |
  \___/  \___/ |    \\//    |

─────────────────────────────────────────────────────────────────────────
Loading  QDistRnd 0.7.0 (Calculate the distance of a q-ary quantum stabilizer code)
by Leonid P. Pryadko (http://faculty.ucr.edu/~leonid) and
   Vadim A. Shabashov (https://sites.google.com/view/vadim-shabashov/).
maintained by:
   Leonid P. Pryadko (http://faculty.ucr.edu/~leonid),
   Vadim A. Shabashov (https://sites.google.com/view/vadim-shabashov/), and
   Valerii K. Kozin (kozin.valera@gmail.com).
Homepage: https://QEC-pages.github.io/qdistrnd
Report issues at https://github.com/QEC-pages/QDistRnd/issues
─────────────────────────────────────────────────────────────────────────
true
gap> TestPackage("qdistrnd");
Architecture: arm-apple-darwin21.3.0-default64-kv7

testing: /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qdistrnd01.\
tst
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd01.tst:25
# Input is:
WriteMTXE("tmp/n5_q3_complex.mtx",3,mat,
        "% The 5-qubit code [[5,1,3]]_3",
        "% Generated from h(x)=1+x^3-x^5-x^6",
        "% Example from the QDistRnd GAP package"   : field:=F);
# Expected output:
File tmp/n5_q3_complex.mtx was created
# But found:
Error, PrintTo: cannot open 'tmp/n5_q3_complex.mtx' for output
########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd01.tst:32
# Input is:
lis:=ReadMTXE("tmp/n5_q3_complex.mtx");;
# Expected output:
# But found:
Error, no method found! For debugging hints type ?Recovery from NoMethodFo\
und
Error, no 1st choice method found for `ReadAll' on 1 arguments
The 1st argument is 'fail' which might point to an earlier problem

########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd01.tst:33
# Input is:
lis[1]; # the field
# Expected output:
GF(3)
# But found:
Error, Variable: 'lis' must have a value
########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd01.tst:35
# Input is:
lis[2]; # converted to `mode=1`
# Expected output:
1
# But found:
Error, Variable: 'lis' must have a value
########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd01.tst:37
# Input is:
Display(lis[3]);
# Expected output:
 1 . . 1 . 2 2 . . .
 . . 1 . . 1 . 2 2 .
 2 . . . 1 . . 1 . 2
 . 2 2 . . . 1 . . 1
# But found:
Error, Variable: 'lis' must have a value
########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd01.tst:44
# Input is:
lisX:=ReadMTXE("matrices/QX80.mtx",0);;
# Expected output:
# But found:
Error, no method found! For debugging hints type ?Recovery from NoMethodFo\
und
Error, no 1st choice method found for `ReadAll' on 1 arguments
The 1st argument is 'fail' which might point to an earlier problem

########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd01.tst:45
# Input is:
GX:=lisX[3];;
# Expected output:
# But found:
Error, Variable: 'lisX' must have a value
########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd01.tst:46
# Input is:
lisZ:=ReadMTXE("matrices/QZ80.mtx",0);;
# Expected output:
# But found:
Error, no method found! For debugging hints type ?Recovery from NoMethodFo\
und
Error, no 1st choice method found for `ReadAll' on 1 arguments
The 1st argument is 'fail' which might point to an earlier problem

########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd01.tst:47
# Input is:
GZ:=lisZ[3];;
# Expected output:
# But found:
Error, Variable: 'lisZ' must have a value
########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd01.tst:48
# Input is:
DistRandCSS(GX,GZ,100,1,2:field:=GF(2));
# Expected output:
5
# But found:
Error, Variable: 'GX' must have a value
########
      35 ms (28 ms GC) and 1.29MB allocated for qdistrnd01.tst
testing: /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qdistrnd02.\
tst
      32 ms (25 ms GC) and 3.71MB allocated for qdistrnd02.tst
testing: /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qdistrnd03.\
tst
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd03.tst:14
# Input is:
    lis:=ReadMTXE("matrices/n5k1A.mtx" );;
# Expected output:
# But found:
Error, no method found! For debugging hints type ?Recovery from NoMethodFo\
und
Error, no 1st choice method found for `ReadAll' on 1 arguments
The 1st argument is 'fail' which might point to an earlier problem

########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd03.tst:15
# Input is:
    Print("field ",lis[1],"\n");
# Expected output:
field GF(7)
# But found:
Error, Variable: 'lis' must have a value
########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd03.tst:17
# Input is:
    dist:=DistRandStab(lis[3],100,0 : field:=lis[1]);
# Expected output:
3
# But found:
Error, Variable: 'lis' must have a value
########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd03.tst:21
# Input is:
    lis:=ReadMTXE("matrices/n5k1.mtx" );;
# Expected output:
# But found:
Error, no method found! For debugging hints type ?Recovery from NoMethodFo\
und
Error, no 1st choice method found for `ReadAll' on 1 arguments
The 1st argument is 'fail' which might point to an earlier problem

########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd03.tst:22
# Input is:
    Print("field ",lis[1],"\n");
# Expected output:
field GF(2)
# But found:
Error, Variable: 'lis' must have a value
########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd03.tst:24
# Input is:
    dist:=DistRandStab(lis[3],100,0,2 : field:=lis[1]);
# Expected output:
3
# But found:
Error, Variable: 'lis' must have a value
########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd03.tst:27
# Input is:
    lis:=ReadMTXE("matrices/n5k1.mtx" : field:= GF(q));;
# Expected output:
# But found:
Error, no method found! For debugging hints type ?Recovery from NoMethodFo\
und
Error, no 1st choice method found for `ReadAll' on 1 arguments
The 1st argument is 'fail' which might point to an earlier problem

########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd03.tst:28
# Input is:
    Print("field ",lis[1],"\n");
# Expected output:
field GF(17)
# But found:
Error, Variable: 'lis' must have a value
########
########> Diff in /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-0.7.0/tst/qd\
istrnd03.tst:30
# Input is:
    dist:=DistRandStab(lis[3],100,0,2 : field:=lis[1]);
# Expected output:
3
# But found:
Error, Variable: 'lis' must have a value
########
      25 ms (25 ms GC) and 122KB allocated for qdistrnd03.tst
-----------------------------------
total        92 ms (78 ms GC) and 5.12MB allocated
             19 failures in 2 of 3 files

#I  Errors detected while testing
```
