### 2022-02-19 8ed111c
Made sure tests can be run from anywhere; modified README.md
