### 2022-02-19 QEC-pages/QDistRnd#9 @LeonidPryadko
The issue is resolved in the branch "ver08".  You can try that branch, or, to run the tests in the branch you have, type
ChangeDirectoryCurrent("absolute_path_to_the_package_QDistRnd");
at the GAP prompt before running the tests.
