### 2022-02-20 QEC-pages/QDistRnd#9 @oscarhiggott
Perfect, thanks! I tried ver08 as well as changing directory with 0.7.0 and both fixed the issue
