### 2022-02-21 QEC-pages/QDistRnd#5 @LeonidPryadko
This issue is fixed.  Explicitly specified primitive polynomial is implemented both in `WriteMTXE` and in `ReadMTXE`, with the latter function doing automatic conversion if necessary.
