### 2022-02-21 #4120 @LeonidPryadko
Two problems have been corrected and the corresponding issues closed on the GitHub page for the program.
