### 2022-02-21 QEC-pages/QDistRnd#13 @oscarhiggott
Specify working directory in docs
From https://github.com/openjournals/joss-reviews/issues/4120

Although I expect most users would work it out anyway, it might be worth adding a sentence at the beginning of Chapter 2 of the docs stating that the examples should be run from the root of the QDistRnd package directory (or alternatively have examples that can run from any directory)
