### 2022-02-21 QEC-pages/QDistRnd#14 @oscarhiggott
LaTeX not rendering in docs
From https://github.com/openjournals/joss-reviews/issues/4120

The LaTeX in the docs doesn't currently render, which is most noticeable in Chapter 3: https://qec-pages.github.io/QDistRnd/doc/chap3.html. It's not important (only worth changing if it's an easy fix) but it would be much easier to read if it did render.

E.g. here:
https://github.com/QEC-pages/QDistRnd/blob/2b8b6da8c3311baf6cb697db8f2b2b46f6a231ce/doc/body.autodoc#L352-L354
