### 2022-02-22 QEC-pages/QDistRnd#14 @LeonidPryadko
The equations require that MathJax be on.  For some reason, it is off by default, but can be enabled by clicking "MathJax on" link near the top of the page (at least this is how it works on Firefox).  Alternatively, there is also a pdf version of the manual at https://github.com/QEC-pages/QDistRnd/releases.
