### 2022-02-22 QEC-pages/QDistRnd#14 @oscarhiggott
Ah sorry I didn't notice the "MathJax on" button! It might be worth linking to the url with MathJax on by default in https://qec-pages.github.io/QDistRnd/ (e.g. https://qec-pages.github.io/QDistRnd/doc/chap0_mj.html instead of https://qec-pages.github.io/QDistRnd/doc/chap0.html) as I imagine most users would prefer that, but closing this issue now
