### 2022-02-22 QEC-pages/QDistRnd#15 @oscarhiggott
maxav docstring links to incorrect section
From https://github.com/openjournals/joss-reviews/issues/4120

I assume this should link to `<Ref Subsect="Subsection_Empirical"/>`

https://github.com/QEC-pages/QDistRnd/blob/2b8b6da8c3311baf6cb697db8f2b2b46f6a231ce/lib/qdistrnd.g#L754-L755

I believe the same link is also incorrect here:
https://github.com/QEC-pages/QDistRnd/blob/2b8b6da8c3311baf6cb697db8f2b2b46f6a231ce/lib/qdistrnd.g#L910-L911
