### 2022-02-22 QEC-pages/QDistRnd#16 @oscarhiggott
Add CONTRIBUTING.md
From https://github.com/openjournals/joss-reviews/issues/4120

The review checklist requires:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Point (2) and I think (3) are covered by your reference to the issue tracker on the package website, but I can't find any contributing guidelines for (1).

The easiest way to address this would be to add a `CONTRIBUTING.md` file to the root of the repository, or add a section in the `README.md` or package website.
