### 2022-02-22 QEC-pages/QDistRnd#17 @oscarhiggott
Pair not specified in MTXE format
From https://github.com/openjournals/joss-reviews/issues/4120

I noticed that the `pair` variable is not recorded in the MTXE file format (unless the user chooses to add it in a comment). The required `type` variable stored in the file does not distinguish between pair values of 0, 1 or 2. It therefore seems to me that a user could easily store a file with `pair=1`, for example, and then accidentally use `ReadMTXE` to load it with `pair=2`, leading to incorrectly permuted columns.

Is there a good reason why you do not want `pair` to be explicitly specified in the MTXE format, such that `ReadMTXE` can by default use the correct settings when loading?
