### 2022-02-22 QEC-pages/QDistRnd#17 @oscarhiggott
I notice that pair=2 isn't supported, but perhaps pair = 0 could still be mistakenly interpreted as pair = 1 or vice versa?
