### 2022-02-22 QEC-pages/QDistRnd#18 @oscarhiggott
Exception from empty comment in MTXE header
From https://github.com/openjournals/joss-reviews/issues/4120

I get the following error when the second line of the MTXE file is an empty comment:

```
gap> lisz:=ReadMTXE("tmp/toric_L_3_Hz.mtx");;
Error, List Element: <list>[2] must have an assigned value in
  line[1][1] = '%' at /Users/oscarhiggott/gap-4.11.1/pkg/QDistRnd-ver08/lib/qdistrnd.g:208 called from
<function "ReadMTXE">( <arguments> )
 called from read-eval loop at *stdin*:6
```

For some reason the default option for `scipy.io.mmwrite` (docs [here](https://docs.scipy.org/doc/scipy/reference/generated/scipy.io.mmwrite.html)) is to have an empty comment on line two, which is why I encountered this issue.

For example this is the `toric_L_3_Hz.mtx` file I was trying to load, created with `scipy.io.mmwrite`:

```
%%MatrixMarket matrix coordinate integer general
%
9 18 36
1 1 1
1 7 1
1 10 1
1 12 1
2 2 1
2 8 1
2 10 1
2 11 1
3 3 1
3 9 1
3 11 1
3 12 1
4 1 1
4 4 1
4 13 1
4 15 1
5 2 1
5 5 1
5 13 1
5 14 1
6 3 1
6 6 1
6 14 1
6 15 1
7 4 1
7 7 1
7 16 1
7 18 1
8 5 1
8 8 1
8 16 1
8 17 1
9 6 1
9 9 1
9 17 1
9 18 1

```
