### 2022-02-22 QEC-pages/QDistRnd#17 @oscarhiggott
To be more concrete, I suppose it would be nice if the `WriteMTXE` function added a comment line by default after the "Field" line, specifying the pair. E.g.:

```
%%MatrixMarket matrix coordinate integer general
% Field: GF(2)
% Pair: 0
```

And then if the `% Pair: 0` line was automatically used (if present) when using `ReadMTXE`. But perhaps there is a reason why you do not want to add this.
