### 2022-02-22 QEC-pages/QDistRnd#17 @LeonidPryadko
Thanks very much for this and other thoughtful comments.
The main reason for not using the pair field is to simplify the format as
much as possible.
To avoid confusion, the best way is to store CSS matrices with pair=0, and
for non-CSS codes to use the complex format;
This should be the default when exchanging matrices with others.
I will add a note to this end on the page describing the format.  -- Leonid

On Tue, Feb 22, 2022 at 10:46 AM oscarhiggott ***@***.***>
wrote:

> To be more concrete, I suppose it would be nice if the WriteMTXE function
> added a comment line by default after the "Field" line, specifying the
> pair. E.g.:
>
> %%MatrixMarket matrix coordinate integer general
> % Field: GF(2)
> % Pair: 0
>
> And then if the % Pair: 0 line was automatically used (if present) when
> using ReadMTXE. But perhaps there is a reason why you do not want to add
> this.
>
