### 2022-02-22 QEC-pages/QDistRnd#17 @oscarhiggott
Great thanks, that makes sense, and a note about this on the page would indeed be helpful
