### 2022-03-01 #4120 @danielskatz
@oscarhiggott - it looks like @LeonidPryadko has addressed 2 of your open issues, though not  https://github.com/QEC-pages/QDistRnd/issues/16 for some reason (or maybe this was addressed too, just not commented upon in the issue?)
