+++ doc/QDistRnd.bib
@@ -280,6 +280,75 @@
   ISSN =	 {0018-9448},
 }
 
+@Article{Cuellar-etal-2020,
+  author =	 {M. P. Cu{\'e}llar and G{\'o}mez-Torrecillas, J.  and
+                  F. J. Lobillo and G. Navarro},
+  title =	 {Genetic algorithms with permutation-based
+                  representation for computing the distance of linear
+                  codes},
+  year =	 2021,
+  journal =	 {Swarm and Evolutionary Computation},
+  volume =	 60,
+  pages =	 100797,
+  eprint =	 {arXiv:2002.12330},
+  doi =		 {10.1016/j.swevo.2020.100797},
+  annote =	 {Finding the minimum distance of linear codes is an
+                  NP-hard problem. Traditionally, this computation has
+                  been addressed by means of the design of algorithms
+                  that find, by a clever exhaustive search, a linear
+                  combination of some generating matrix rows that
+                  provides a codeword with minimum weight. Therefore,
+                  as the dimension of the code or the size of the
+                  underlying finite field increase, so it does
+                  exponentially the run time. In this work, we prove
+                  that, given a generating matrix, there exists a
+                  column permutation which leads to a reduced row
+                  echelon form containing a row whose weight is the
+                  code distance. This result enables the use of
+                  permutations as representation scheme, in contrast
+                  to the usual discrete representation, which makes
+                  the search of the optimum polynomial time dependent
+                  from the base field. In particular, we have
+                  implemented genetic and CHC algorithms using this
+                  representation as a proof of concept. Experimental
+                  results have been carried out employing codes over
+                  fields with two and eight elements, which suggests
+                  that evolutionary algorithms with our proposed
+                  permutation encoding are competitive with regard to
+                  existing methods in the literature. As a by-product,
+                  we have found and amended some inaccuracies in the
+                  MAGMA Computational Algebra System concerning the
+                  stored distances of some linear codes.}
+}
+
+@article{Dumer-Kovalev-Pryadko-bnd-2015,
+  title =	 {Thresholds for correcting errors, erasures, and
+                  faulty syndrome measurements in degenerate quantum
+                  codes},
+  author =	 {Dumer, I. and Kovalev, A. A. and Pryadko, L. P.},
+  journal =	 {Phys. Rev. Lett.},
+  volume =	 115,
+  pages =	 050502,
+  numpages =	 5,
+  year =	 2015,
+  publisher =	 {American Physical Society},
+  doi =		 {10.1103/PhysRevLett.115.050502},
+  url =
+                  {http://link.aps.org/doi/10.1103/PhysRevLett.115.050502},
+  eprint =	 {1412.6172},
+  annote =	 {We suggest a technique for constructing lower
+                  (existence) bounds for the fault-tolerant threshold
+                  to scalable quantum computation applicable to
+                  degenerate quantum codes with sublinear distance
+                  scaling. We give explicit analytic expressions
+                  combining probabilities of erasures, depolarizing
+                  errors, and phenomenological syndrome measurement
+                  errors for quantum LDPC codes with logarithmic or
+                  larger distances. These threshold estimates are
+                  parametrically better than the existing analytical
+                  bound based on percolation.  }
+}
+
 @ARTICLE{Dumer-Kovalev-Pryadko-IEEE-2017,
   author =	 {I. Dumer and A. A. Kovalev and L. P. Pryadko},
   journal =	 {IEEE Trans. Inf. Th.},
