### 2022-03-01 QEC-pages/QDistRnd#17 @LeonidPryadko
Documentation updated, should be no confusion now.
