+++ CONTRIBUTING.md
@@ -1,0 +1,38 @@
+# Contributing 
+
+We invite everyone to contribute by submitting patches, pull requests,
+bug reports, and code reviews. 
+
+##  Issue reporting and code contributions
+
+* Before you report an issue, or wish to add functionality, please try
+  and check to see if there are existing issues or pull requests. We
+  do not want you wasting your time duplicating somebody else's work.
+	
+* For substantial changes it is also advisable to contact us before
+  you start work to discuss your ideas.
+  
+* You should be prepared to wait until your pull request or patch has
+  been discussed and authorized prior to its inclusion.  We might also
+  ask for you to adapt your changes to make them suitable for
+  inclusion.
+  
+* To help increase the chance of your pull request being accepted:
+
+  - Run the tests.
+  
+  - Update the documentation, tests, examples, guides, and whatever
+    else is affected by your contribution.
+	
+  - Use appropriate code formatting for your code.
+
+* Try to follow *the Campsite Rule:* leave the codebase in better
+  condition than you found it. Please clean up any messes that you
+  find, and don't leave behind new messes for the next contributor.
+
+## How do I make a contribution?
+
+Never made an open source contribution before? Wondering how
+contributions work in our project? Please follow the guidelines at
+(https://github.com/gap-system/gap/blob/master/CONTRIBUTING.md).
+
