### 2022-03-01 QEC-pages/QDistRnd#16 @LeonidPryadko
Added a CONTRIBUTING.md file.
