### 2022-03-08 QEC-pages/QDistRnd#18 @LeonidPryadko
I think it can be closed but I was hoping for a confirmation from
@oscarhiggot

Leonid

On Mon, Mar 7, 2022 at 5:25 AM Daniel S. Katz ***@***.***>
wrote:

> Should this issue now be closed?
>
