### 2022-03-08 QEC-pages/QDistRnd#18 @oscarhiggott
I just tried it and can confirm the latest version has fixed the bug, so will close the issue now, thanks!
