### 2022-03-08 QEC-pages/QDistRnd#18 @LeonidPryadko
Thank you very much! -- Leonid

On Tue, Mar 8, 2022 at 11:30 AM oscarhiggott ***@***.***>
wrote:

> Closed #18 <https://github.com/QEC-pages/QDistRnd/issues/18>.
>
