### 2022-03-16 #4120 @pan-pavel
Hello Daniel,

Sorry for the late response! Yes, I think we can now proceed to publication.

Best Regards,
Pavel

> 14 марта 2022 г., в 17:52, Daniel S. Katz ***@***.***> написал(а):
>
> ﻿
> 👋 @pan-pavel - Can you let us know if you are ok with this now proceeding to publication?
>
