### 2022-03-16 #4120 @danielskatz
@LeonidPryadko - I've suggested some minor changes to the paper in https://github.com/QEC-pages/QDistRnd/pull/20 - If you can merge this, we can move forward to acceptance and publication.  Let me know if you disagree with any of these changes.
