### 2022-03-16 #4120 @LeonidPryadko
The changes look fine.  I am about to merge them, and will make a couple of
additional small edits to the manuscript in the next half an hour or so.

Leonid

On Wed, Mar 16, 2022 at 1:47 PM Daniel S. Katz ***@***.***>
wrote:

> @LeonidPryadko <https://github.com/LeonidPryadko> - I've suggested some
> minor changes to the paper in QEC-pages/QDistRnd#20
> <https://github.com/QEC-pages/QDistRnd/pull/20> - If you can merge this,
> we can move forward to acceptance and publication. Let me know if you
> disagree with any of these changes.
>
