### 2022-03-16 6f6a6ea
Update paper.md
A small correction
