### 2022-03-17 #4120 @LeonidPryadko
I am done with any changes to the paper.md file.

Regarding the items you requested, do you have a more detailed set of
instructions, please?
I do not really understand what "tagged release" means, nor where exactly
the version needs to be listed, or where to get the DOI from.

Thanks,

Leonid

On Wed, Mar 16, 2022 at 4:17 PM Daniel S. Katz ***@***.***>
wrote:

> When you are done, please:
>
>    - Make a tagged release of your software, and list the version tag of
>    the archived version here.
>    - Archive the reviewed software in Zenodo or a similar service (e.g.,
>    figshare, an institutional repository)
>    - Check the archival deposit (e.g., in Zenodo) has the correct
>    metadata. This includes the title (should match the paper title) and author
>    list (make sure the list is correct and people who only made a small fix
>    are not on it). You may also add the authors' ORCID.
>    - Please list the DOI of the archived version here.
>
> I can then move forward with accepting the submission.
>
