### 2022-03-17 #4120 @danielskatz
A tagged release is really a snapshot of the repository at a given time, with a label (typically a version number) - see https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content

And what JOSS needs is for that snapshot to be deposited in an archive that will preserve it and gives it a DOI, which is most easily done via Zenodo for GitHub repositories, following the instructions in https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content (though note that you can also use other archival services if you wish)

To summarize, the simplest path is to
1. Follow https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content to create a link/hook from your GitHub repository to Zenodo
2. See https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content for how to create a release on GitHub, which will then trigger an archive being stored on Zenodo
3. Go into Zenodo and update the metadata for that archive to make sure the title and authors match those of the paper (rather than those of the repository). See the section labeled "How can I edit the metadata of a published record?" in https://help.zenodo.org for how to do this.
