+++ README.md
@@ -1,267 +1,70 @@
-# GitHubPagesForGAP
+# gap-package-qdistrnd
+The `GAP` package for calculating the distance of a $q$-ary quantum
+stabilizer code.
 
-This repository can be used to quickly set up a website hosted by
-[GitHub](https://github.com/) for GAP packages using a GitHub repository.
-Specifically, this uses [GitHub pages](https://pages.github.com/)
-by adding a `gh-pages` branch to your package repository which
-contains data generated from the `PackageInfo.g` file of your package.
+## Overview
+The `GAP` package `QDistRnd` implements a probabilistic algorithm for
+finding the distance of a $q$-ary quantum low-density parity-check
+code linear over a finite field GF($q$).  An empirical convergence
+criterion is provided to estimate the probability that a minimum
+weight codeword has actually been found.  Versions of the routines for
+CSS and regular stabilizer codes are given.
 
-## Initial setup
+In addition, a format for storing matrices associated with $q$-ary
+quantum codes is introduced and implemented via provided import/export
+routines.  The format is based on the well established MaTrix market
+eXchange (MTX) Coordinate format developed at NIST, and is designed
+for full backward compatibility with this format.  Thus, the files are
+readable by any software package which supports MTX.
 
-The easiest way to do this is to run the `setup-gh-pages` shell script
-provided in the [GitHubPagesForGAP]() from within a git clone of your
-package's GitHub repository.
+## Requirements 
+The package `QDistRnd` requires the `Guava` package to run; `GAPDoc`
+and `AutoDoc` are required to generate the documentation (see the file
+`PackageInfo.g` for versions required).  All of these packages are
+included with `GAP` version 4.11; this or later version of `GAP` is
+strongly recommended.
 
-In case this does not work, or if you want to really know what's going
-on, you can also follow the manual instructions described after the fold.
+## Installation
+As of February 2022, the `QDistRnd` package is not distributed with a
+standard installation of `GAP`. 
 
-------
+To install the package permanently, download the latest released version from 
+[releases](https://github.com/QEC-pages/QDistRnd/releases/) 
+and unpack it in the `pkg` directory of one of your `GAP` root
+directories.  After installation, the package can be loaded at the
+`GAP` prompt by typing
 
-The following instructions assume you do not already have a `gh-pages`
-branch in your repository. If you do have one, you should delete it before
-following these instructions.
+gap> `LoadPackage("QDistRnd");`
 
-1. Go into your clone of your package repository.
+Alternatively, if you just want to try it, you can unpack the package
+anywhere and type at the `GAP` prompt 
 
-2. Setup a `gh-pages` branch in a `gh-pages` subdirectory.
+gap> `SetPackagePath("QDistRnd","absolute_path_to_the_package_QDistRnd" );`
 
-   Users with a recent enough git version (recommended is >= 2.7.0)
-   can do this using a "worktree", via the following commands:
+After that you can load the package as you would do normally.
 
-   ```sh
-   # Add a new remote pointing to the GitHubPagesForGAP repository
-   git remote add -f gh-gap https://github.com/gap-system/GitHubPagesForGAP
+## Testing
 
-   # Create a fresh gh-pages branch from the new remote
-   git branch gh-pages gh-gap/gh-pages --no-track
+After installation, basic tests of the package (most of the examples
+listed in the package manual) can by performed by running
 
-   # Create a new worktree and change into it
-   git worktree add gh-pages gh-pages
-   cd gh-pages
-   ```
+gap> `TestPackage("qdistrnd");`
[…]
