### 2022-03-22 #4120 @LeonidPryadko
I created a release on Zenodo
https://zenodo.org/record/6374559

I also tried to follow the instructions on the release page and created a
.zenodo.json file in the repository.
However, after re-releasing the package, the Zenodo release failed, with a
cryptic message.
Any suggestions?
Specifically, is there a way to verify the validity of the `.zenodo.json`
file without generating a release?

Thanks,

Leonid

On Mon, Mar 21, 2022 at 8:11 AM Daniel S. Katz ***@***.***>
wrote:

> @LeonidPryadko <https://github.com/LeonidPryadko> - did my
> comment/explanation help?
>
