+++ CITATION.cff
@@ -9,7 +9,7 @@ authors:
   - family-names: Kozin
     given-names: "Valerii K."
 title: "QDistRnd: Calculate the distance of a q-ary quantum stabilizer code"
-version: v0.8.4
+version: v0.8.5
 date-released: 2022-03-21
 repository-code: "https://github.com/QEC-pages/QDistRnd"
 url: "https://QEC-pages.github.io/QDistRnd/"
