### 2022-03-22 #4120 @LeonidPryadko
I just corrected the title and edited the description a bit in the last
version released,
https://doi.org/10.5281/zenodo.6374733

Please let me know if any other actions are needed.

Thank you,

Leonid

On Tue, Mar 22, 2022 at 5:28 AM Daniel S. Katz ***@***.***>
wrote:

> At this point, I suggest we use
> https://zenodo.org/record/6374733#.YjnAPC-B1_4 (
> https://doi.org/10.5281/zenodo.6374733), and you just updated the
> metadata in this deposit so that the title matches the title of the paper.
> See the section labeled "How can I edit the metadata of a published
> record?" in https://help.zenodo.org/ for how to do this.
>
