### 2022-03-22 #4120 @xuanxu
@danielskatz it seems there's a reference in the .bib file missing the author
