### 2022-03-22 #4120 @danielskatz
ok, thanks - I'll look into it.

But why don't we catch this in generate pdf?  there shouldn't be any new error in recommend-accept, I don't think
