+++ joss/QDistRnd.bib
@@ -550,6 +550,7 @@
 }
 
 @misc{nist-mm-format,
+  author = {{National Institute of Standards and Technology}},
   howpublished =  {online},
   note =          {Accessed on May 27, 2019},
   title =         {Matrix Market exchange formats},
