### 2022-03-22 #4120 @LeonidPryadko
I just merged it.  Please let me know if I should also generate a new
release.

Leonid

On Tue, Mar 22, 2022 at 12:05 PM Daniel S. Katz ***@***.***>
wrote:

> @LeonidPryadko <https://github.com/LeonidPryadko> - I think that if you
> can merge QEC-pages/QDistRnd#21
> <https://github.com/QEC-pages/QDistRnd/pull/21>, it will let us progress
>
