### 2022-10-11 #4120 @LeonidPryadko
Ok, this is a corrected version
https://zenodo.org/record/6374733
with all version numbers up to date.

Thanks for the help!
Anything else I need to do?

Best,

Leonid

On Mon, Mar 21, 2022 at 9:26 PM Leonid Pryadko ***@***.***>
wrote:

> I created a release on Zenodo
> https://zenodo.org/record/6374559
>
> I also tried to follow the instructions on the release page and created a
> .zenodo.json file in the repository.
> However, after re-releasing the package, the Zenodo release failed, with a
> cryptic message.
> Any suggestions?
> Specifically, is there a way to verify the validity of the `.zenodo.json`
> file without generating a release?
>
> Thanks,
>
> Leonid
>
> On Mon, Mar 21, 2022 at 8:11 AM Daniel S. Katz ***@***.***>
> wrote:
>
>> @LeonidPryadko <https://github.com/LeonidPryadko> - did my
>> comment/explanation help?
>>
>> —
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/4120#issuecomment-1074020349>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/APF4LTXTWJGQWPQX272X7VTVBCGQTANCNFSM5NJ2H2YQ>
>> .
>> You are receiving this because you were mentioned.Message ID:
>> ***@***.***>
>>
>
